# apps/openapi.yaml
openapi: 3.1.0
info:
  title: Quant Trading System API
  version: "1.2.0"
  description: >
    OpenAPI specification for the current quant-trading-system project.
    It is organized around the real backend domain model:
    Strategy -> StrategyRun -> Signal / Transaction / PortfolioSnapshot,
    plus market-data and broker-trading integrations.

    Notes:
    - Operations marked with x-implementation-status: implemented are already reflected by the current FastAPI codebase.
    - Operations marked with x-implementation-status: planned are aligned with existing service-layer and database structure,
      and are intended to guide the next API/router implementation pass.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/agent/research/category-studies/validate:
    post:
      tags: [Agent Integration]
      summary: Validate a generated engine-category draft and first adaptive round
      operationId: validateAgentCategoryStudy
      x-implementation-status: implemented
      security:
        - AgentServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryStudyValidationRequest"
      responses:
        "200":
          description: Normalized specification and bounded trial count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryStudyValidation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/agent/research/category-studies:
    post:
      tags: [Agent Integration]
      summary: Create a persistent adaptive category study after experiment approval
      operationId: createAgentCategoryStudy
      x-implementation-status: implemented
      security:
        - AgentServiceToken: []
      parameters:
        - $ref: "#/components/parameters/RequiredIdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryStudyCreate"
      responses:
        "201":
          description: Experiment created or replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchExperiment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/agent/research/experiments/{experiment_id}/rounds:
    post:
      tags: [Agent Integration]
      summary: Submit the next bounded adaptive candidate round
      operationId: submitAgentExperimentRound
      security: [{AgentServiceToken: []}]
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
        - $ref: "#/components/parameters/RequiredIdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExperimentRoundSubmit"}
      responses:
        "200": {description: Updated experiment, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchExperiment"}}}}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
  /api/agent/research/experiments/{experiment_id}/controller-failure:
    post:
      tags: [Agent Integration]
      summary: Terminate an adaptive experiment when its Agent controller cannot continue
      operationId: failAgentExperimentController
      security: [{AgentServiceToken: []}]
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
        - $ref: "#/components/parameters/RequiredIdempotencyKeyHeader"
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ControllerFailure"}}}
      responses:
        "200": {description: Terminal experiment, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchExperiment"}}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/agent/research/experiments/{experiment_id}/candidates/promote:
    post:
      tags: [Agent Integration]
      summary: Save up to five Pareto rank 1-2 candidates as visible drafts
      operationId: promoteAgentExperimentCandidates
      security: [{AgentServiceToken: []}]
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
        - $ref: "#/components/parameters/RequiredIdempotencyKeyHeader"
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/CandidatePromotion"}}}
      responses:
        "200": {description: Promoted draft strategy IDs}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
  /api/agent/research/experiments/{experiment_id}/cancel:
    post:
      tags: [Agent Integration]
      summary: Request cancellation of an experiment
      operationId: cancelAgentExperiment
      x-implementation-status: implemented
      security:
        - AgentServiceToken: []
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      responses:
        "200":
          description: Current experiment state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchExperiment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/agent/research/experiments/{experiment_id}/usage:
    post:
      tags: [Agent Integration]
      summary: Synchronize absolute AgentOps token usage and enforce the experiment stop policy
      operationId: updateAgentExperimentTokenUsage
      x-implementation-status: implemented
      security:
        - AgentServiceToken: []
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentTokenUsageUpdate"
      responses:
        "200":
          description: Current experiment state after idempotent absolute usage synchronization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchExperiment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/agent/research/experiments/{experiment_id}/report/retry:
    post:
      tags: [Agent Integration]
      summary: Idempotently retry effectiveness-study report generation
      operationId: retryAgentEffectivenessReport
      security: [{AgentServiceToken: []}]
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      responses:
        "200": {description: Updated parent experiment, content: {application/json: {schema: {$ref: "#/components/schemas/ResearchExperiment"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
...
  /api/research/experiments/{experiment_id}:
    get:
      tags: [Research]
      summary: Get a research experiment
      operationId: getResearchExperiment
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      responses:
        "200":
          description: Research experiment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchExperiment"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/research/experiments/{experiment_id}/trials:
    get:
      tags: [Research]
      summary: List stable ordered experiment trials
      operationId: listResearchExperimentTrials
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      responses:
        "200":
          description: Experiment trials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExperimentTrial"
  /api/research/experiments/{experiment_id}/backtests/{run_id}:
    delete:
      tags: [Research]
      summary: Delete one terminal research backtest while retaining experiment evidence
      operationId: deleteResearchBacktest
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "200":
          description: Run artifacts deleted; the owning trial or candidate keeps a deletion tombstone
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BacktestDeleteResult"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/research/experiments/{experiment_id}/report:
    get:
      tags: [Research]
      summary: Get the deterministic robustness report
      operationId: getResearchExperimentReport
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
      responses:
        "200":
          description: Robustness report linked to underlying trials and backtests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreeFormObject"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/research/experiments/{experiment_id}/children:
    get:
      tags: [Research]
      summary: List child experiments of a parent effectiveness study
      operationId: listResearchExperimentChildren
      parameters: [{$ref: "#/components/parameters/ExperimentIdPath"}]
      responses:
        "200": {description: Ordered child experiments, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ResearchExperiment"}}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/research/experiments/{experiment_id}/report-artifacts/{artifact_kind}:
    get:
      tags: [Research]
      summary: Download a generated JSON, Markdown, or PDF report artifact
      operationId: downloadResearchReportArtifact
      parameters:
        - $ref: "#/components/parameters/ExperimentIdPath"
        - in: path
          name: artifact_kind
          required: true
          schema: {type: string, enum: [json, markdownZh, markdownEn, pdfZh, pdfEn]}
      responses:
        "200": {description: Generated report artifact}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /api/research/experiments/{experiment_id}/rounds:
    get:
      tags: [Research]
      summary: List adaptive rounds; legacy experiments return an empty list
      operationId: listResearchExperimentRounds
      parameters: [{$ref: "#/components/parameters/ExperimentIdPath"}]
      responses:
        "200": {description: Adaptive rounds, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ExperimentRound"}}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/research/experiments/{experiment_id}/candidates:
    get:
      tags: [Research]
      summary: List adaptive candidates and Pareto ranks; legacy experiments return an empty list
      operationId: listResearchExperimentCandidates
      parameters: [{$ref: "#/components/parameters/ExperimentIdPath"}]
      responses:
        "200": {description: Adaptive candidates, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/ExperimentCandidate"}}}}}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    AgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque service token
  parameters:
...
    RequiredIdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ExperimentIdPath:
      in: path
      name: experiment_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Conflict:
      description: Conflict or duplicate request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation or normalization error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Agent service token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Required worker or data dependency is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    FreeFormObject:
      type: object
      additionalProperties: true
...
    CategoryStudyValidationRequest:
      type: object
      required: [workflowRunId, goal, strategyType, strategy, name, hypothesis]
      properties:
        workflowRunId: {type: string}
        goal: {type: string}
        strategyType: {type: string, enum: [trend, mean_reversion, momentum_breakout, island_reversal, double_bottom, head_shoulders_bottom, rounded_bottom, v_reversal, support_resistance]}
        strategy: {$ref: "#/components/schemas/FreeFormObject"}
        name: {type: string}
        hypothesis: {type: string}
        basketId: {type: [string, "null"], format: uuid}
        symbols: {type: array, items: {type: string}}
        universePolicy: {oneOf: [{$ref: "#/components/schemas/PointInTimeUniversePolicy"}, {type: "null"}]}
        validationProtocol: {oneOf: [{$ref: "#/components/schemas/SupportResistanceValidationProtocol"}, {type: "null"}]}
        inSample: {$ref: "#/components/schemas/DateWindow"}
        outOfSample: {$ref: "#/components/schemas/DateWindow"}
        costScenarios: {type: array, minItems: 2, maxItems: 5, items: {$ref: "#/components/schemas/CostScenario"}}
        initialCash: {type: number, exclusiveMinimum: 0}
        benchmarkSymbol: {type: [string, "null"]}
        stopPolicy: {oneOf: [{$ref: "#/components/schemas/ExperimentStopPolicy"}, {type: "null"}]}
        searchPolicy: {$ref: "#/components/schemas/FreeFormObject"}
        initialCandidates: {type: array, minItems: 1, maxItems: 5, items: {$ref: "#/components/schemas/FreeFormObject"}}
    CategoryStudyValidation:
      type: object
      required: [valid, normalizedStrategy, normalizedSpec, firstRoundTrialCount, maximumTrialCount, universeSymbols, universeSummary, proposalHash]
      properties:
        valid: {type: boolean}
        normalizedStrategy: {$ref: "#/components/schemas/StrategyCreate"}
        normalizedSpec: {$ref: "#/components/schemas/FreeFormObject"}
        firstRoundTrialCount: {type: integer, maximum: 200}
        maximumTrialCount: {type: integer, maximum: 200}
        universeSymbols: {type: array, items: {type: string}}
        universeSummary: {$ref: "#/components/schemas/FreeFormObject"}
        proposalHash: {type: string}
        warnings: {type: array, items: {type: string}}
    CategoryStudyCreate:
      type: object
      required: [workflowRunId, spec]
      properties:
        workflowRunId: {type: string}
        spec: {$ref: "#/components/schemas/FreeFormObject"}
...
    ResearchExperiment:
      type: object
      required: [id, parentExperimentId, studyKind, workflowRunId, status, spec, runManifest, progress, report]
      properties:
        id: {type: string, format: uuid}
        parentExperimentId: {type: [string, "null"], format: uuid}
        studyKind: {type: string}
        workflowRunId: {type: string}
        status:
          type: string
          enum: [queued, running, waiting_agent, completed, partially_failed, failed, cancel_requested, cancelled, data_changed]
        spec: {$ref: "#/components/schemas/FreeFormObject"}
        runManifest: {$ref: "#/components/schemas/FreeFormObject"}
        progress: {$ref: "#/components/schemas/FreeFormObject"}
        report: {$ref: "#/components/schemas/FreeFormObject"}
        errorCode: {type: [string, "null"]}
        errorMessage: {type: [string, "null"]}
        startedAt: {type: [string, "null"], format: date-time}
        finishedAt: {type: [string, "null"], format: date-time}
        createdAt: {type: [string, "null"], format: date-time}
        updatedAt: {type: [string, "null"], format: date-time}
...